refactoring test framework utils
